compilable - hotfix

+++ utility/Util.c
@@ -71,8 +71,8 @@ void DrawCircle(SDL_Surface* screen, int centerX, int centerY, int radius, int w
 		double t = 0;
 		while (t < 360)
 		{
-			int x = r * cos(t * 3.14159265359 / 1.8E2) + centerX;
-			int y = r * sin(t * 3.14159265359 / 1.8E2) + centerY;
+			double x = r * cos(t * 3.14159265359 / 1.8E2) + centerX;
+			double y = r * sin(t * 3.14159265359 / 1.8E2) + centerY;
 			Uint32 color = r < radius - width ? fillColor : outlineColor;
 			DrawPixel(screen, x, y, color);
 			t++;
